+++ test/test_helper.exs
@@ -23,19 +23,19 @@ defmodule Joken.TestPoison do
     nil
   end
 
-  def validate_claim(:exp, payload) do
-    Joken.Config.validate_time_claim(payload, :exp, "Token expired", fn(expires_at, now) -> expires_at > now end)
+  def validate_claim(:exp, payload, _) do
+    Joken.Helpers.validate_time_claim(payload, :exp, "Token expired", fn(expires_at, now) -> expires_at > now end)
   end
 
-  def validate_claim(:nbf, payload) do
-    Joken.Config.validate_time_claim(payload, :nbf, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
+  def validate_claim(:nbf, payload, _) do
+    Joken.Helpers.validate_time_claim(payload, :nbf, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
   end
 
-  def validate_claim(:iat, payload) do
-    Joken.Config.validate_time_claim(payload, :iat, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
+  def validate_claim(:iat, payload, _) do
+    Joken.Helpers.validate_time_claim(payload, :iat, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
   end
 
-  def validate_claim(_, _) do
+  def validate_claim(_, _, _) do
     :ok
   end
 end
@@ -64,19 +64,19 @@ defmodule Joken.TestJsx do
     nil
   end
 
-  def validate_claim(:exp, payload) do
-    Joken.Config.validate_time_claim(payload, :exp, "Token expired", fn(expires_at, now) -> expires_at > now end)
+  def validate_claim(:exp, payload, _) do
+    Joken.Helpers.validate_time_claim(payload, :exp, "Token expired", fn(expires_at, now) -> expires_at > now end)
   end
 
-  def validate_claim(:nbf, payload) do
-    Joken.Config.validate_time_claim(payload, :nbf, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
+  def validate_claim(:nbf, payload, _) do
+    Joken.Helpers.validate_time_claim(payload, :nbf, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
   end
 
-  def validate_claim(:iat, payload) do
-    Joken.Config.validate_time_claim(payload, :iat, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
+  def validate_claim(:iat, payload, _) do
+    Joken.Helpers.validate_time_claim(payload, :iat, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
   end
 
-  def validate_claim(_, _) do
+  def validate_claim(_, _, _) do
     :ok
   end
 end