+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.15.0-dev
+* Enhancements
+  * Add options parameter to `Joken.Config.validate_claim`
+  * Moved validation helpers functions to `Joken.Helpers`
+
 # v0.14.1
 
   * Enhancements